refactor: style, component 관심사 분리(코드분리)
