+++ src/Auth/AuthForm/AuthForm.jsx
@@ -1,19 +1,20 @@
-import styled from 'styled-components';
-
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-
-import { ACCESS_TYPE, initialUserInfo } from '../constants';
 
 import {
   validationUserLogin,
   validationUserRegister
-} from '../utils/functions';
+} from '../../utils/functions';
 
-import AuthButton from '../components/Auth/AuthButton';
-import AuthContent from "../components/Auth/AuthContent";
-import InputWithLabel from "../components/Auth/InputWithLabel";
-import { axios } from '../lib/axios';
+import { axios } from '../../lib/axios';
+
+import { ACCESS_TYPE, initialUserInfo } from '../../constants';
+
+import AuthButton from '../../components/Auth/AuthButton/AuthButton';
+import AuthContent from "../../components/Auth/AuthContent/AuthContent";
+import InputWithLabel from "../../components/Auth/InputWithLabel/InputWithLabel";
+
+import { StyledInputWrapper } from './AuthForm.style';
 
 export default function AuthForm({ accessType }) {
   const navigate = useNavigate();
@@ -102,9 +103,3 @@ export default function AuthForm({ accessType }) {
     </AuthContent>
   );
 };
-
-const StyledInputWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-`;
