+++ src/Auth/AuthForm/AuthForm.style.jsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const StyledInputWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+`;
