+++ src/Todo/Todo.style.jsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const StyledTodoForm = styled.div`
+  display: flex;
+  background-color: white;
+  justify-content: center;
+  width: 600px;
+  height: 500px;
+  border-radius: 20px;
+`;
