+++ src/Todo/TodoForm/TodoForm.style.jsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+
+export const StyledWrapper = styled.div`
+  display: flex;
+  flex-direction: row;
+  gap: 5px;
+  width: 100%;
+  justify-content: space-around;
+`;
