+++ src/Todo/TodoItem/TodoItem.style.jsx
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+
+export const StyledLi = styled.li`
+  list-style: none;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 300px;
+`;
+
+export const StyledButtonWrapper = styled.div`
+  display: flex;
+  gap: 2px;
+`;
+
+export const StyledTodo = styled.div`
+  width: 200px;
+  word-break:break-all;
+`;
+
+export const StyledInput = styled.input`
+  
+`;
