+++ src/Todo/TodoList/TodoList.style.jsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const StyledUl = styled.ul`
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+`;
