+++ src/components/Auth/AuthButton/AuthButton.jsx
@@ -1,0 +1,11 @@
+import { StyledWrapper } from "./AuthButton.style";
+
+export default function AuthButton({ children, onClick, ...rest }) {
+  return (
+    <StyledWrapper onClick={onClick} {...rest}>
+      {children}
+    </StyledWrapper>
+  );
+};
+
+
