+++ src/components/Auth/AuthContent/AuthContent.jsx
@@ -1,0 +1,10 @@
+import { StyledTitle } from "./AuthContent.style";
+
+export default function AuthContent({ title, children }) {
+  return (
+    <div>
+      <StyledTitle>{title}</StyledTitle>
+      {children}
+    </div>
+  );
+};
