+++ src/components/Auth/AuthContent/AuthContent.style.jsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const StyledTitle = styled.div`
+    font-size: 2.5rem;
+    font-weight: 500;
+    margin-bottom: 1rem;
+    text-align: center;
+`;
