+++ src/components/Auth/AuthLogo/AuthLogo.jsx
@@ -1,0 +1,9 @@
+import { StyledLogo, StyledLogoWrapper } from "./AuthLogo.style";
+
+export default function AuthLogo() {
+  return (
+    <StyledLogoWrapper>
+      <StyledLogo>JH-TODO-LIST</StyledLogo>
+    </StyledLogoWrapper>
+  );
+};
