+++ src/components/Auth/AuthLogo/AuthLogo.style.jsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const StyledLogoWrapper = styled.div`
+    height: 5rem;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+`;
+
+export const StyledLogo = styled.div`
+    font-size: 3.6rem;
+    letter-spacing: 5px;
+    text-decoration: none;
+`;
