+++ src/components/Auth/AuthWrapper/AuthWrapper.style.jsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const StyledPositioner = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 70px;
+`;
+
+export const StyledContents = styled.div`
+  padding: 2rem;
+`;
