+++ src/components/Auth/InputWithLabel/InputWithLabel.jsx
@@ -1,0 +1,12 @@
+import { StyledInput, StyledLabel, StyledWrapper } from "./InputWithLabel.style";
+
+export default function InputWithLabel({ label, ...rest }) {
+  return (
+    <StyledWrapper>
+      <StyledLabel htmlFor={label}>
+        {label}
+      </StyledLabel>
+      <StyledInput id={label} {...rest} />
+    </StyledWrapper>
+  );
+};
