+++ src/components/Auth/InputWithLabel/InputWithLabel.style.jsx
@@ -1,0 +1,22 @@
+import styled from 'styled-components';
+
+export const StyledWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
+
+export const StyledLabel = styled.label`
+  font-size: 1rem;  
+  margin-bottom: 0.25rem;
+`;
+
+export const StyledInput = styled.input`
+    width: 350px;
+    border: 1px solid black;
+    outline: none;
+    border-radius: 0px;
+    line-height: 2.5rem;
+    font-size: 1.2rem;
+    padding-left: 0.5rem;
+    padding-right: 0.5rem;
+`;
