+++ src/components/Todo/TodoButton/TodoButton.jsx
@@ -1,0 +1,7 @@
+import { StyledButton } from "./TodoButton.style";
+
+export default function TodoButton({ children, onClick, ...rest }) {
+  return (
+    <StyledButton {...rest} onClick={onClick}>{children}</StyledButton>
+  );
+}
