+++ src/components/Todo/TodoInput/TodoInput.jsx
@@ -1,0 +1,7 @@
+import { StyledInput } from "./TodoInput.style";
+
+export default function TodoInput({ ...rest }) {
+  return (
+    <StyledInput {...rest} />
+  );
+};
