+++ src/components/Todo/TodoInput/TodoInput.style.jsx
@@ -1,12 +1,6 @@
 import styled from 'styled-components';
 
-export default function TodoInput({ ...rest }) {
-  return (
-    <StyledInput {...rest} />
-  );
-};
-
-const StyledInput = styled.input`
+export const StyledInput = styled.input`
   width: 80%;
   border: none;
   border-bottom: 1px solid black;
