+++ src/components/Todo/TodoTitle/TodoTitle.jsx
@@ -1,0 +1,9 @@
+import { StyledTitle, StyledTitleWrapper } from "./TodoTitle.style";
+
+export default function TodoTitle() {
+  return (
+    <StyledTitleWrapper>
+      <StyledTitle>Todo</StyledTitle>
+    </StyledTitleWrapper>
+  );
+};
