+++ src/components/Todo/TodoWrapper/TodoWrapper.jsx
@@ -1,0 +1,14 @@
+import TodoTitle from "../TodoTitle/TodoTitle";
+
+import { StyledContents, StyledPositioner } from "./TodoWrapper.style";
+
+export default function TodoWrapper({ children }) {
+  return (
+    <StyledPositioner>
+      <TodoTitle />
+      <StyledContents>
+        {children}
+      </StyledContents>
+    </StyledPositioner>
+  );
+};
