+++ src/components/Todo/TodoWrapper/TodoWrapper.style.jsx
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const StyledPositioner = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 50px;
+  padding: 2rem;
+  width: 100%;
+  height: 100%;
+`;
+
+export const StyledContents = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+`;
